Move concurrent unit test
