+++ platform-sdk/consensus-concurrent/build.gradle.kts
@@ -2,6 +2,7 @@
 plugins {
     id("org.hiero.gradle.module.library")
     id("org.hiero.gradle.feature.publish-artifactregistry")
+    id("org.hiero.gradle.feature.test-fixtures")
 }
 
 description = "Consensus Concurrent"
@@ -9,6 +10,13 @@ description = "Consensus Concurrent"
 mainModuleInfo { annotationProcessor("com.swirlds.config.processor") }
 
 testModuleInfo {
+    requires("com.swirlds.common.test.fixtures")
     requires("com.swirlds.config.extensions.test.fixtures")
+    requires("org.hiero.base.utility.test.fixtures")
+    requires("org.hiero.consensus.concurrent.test.fixtures")
+    requires("org.hiero.consensus.metrics")
+    requires("org.assertj.core")
     requires("org.junit.jupiter.api")
+    requires("org.junit.jupiter.params")
+    requires("org.mockito")
 }
