+++ platform-sdk/consensus-concurrent/src/test/java/org/hiero/consensus/concurrent/pool/StandardWorkGroupTest.java
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: Apache-2.0
-package com.swirlds.common.threading.pool;
+package org.hiero.consensus.concurrent.pool;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.hiero.consensus.concurrent.manager.AdHocThreadManager.getStaticThreadManager;
@@ -10,7 +10,6 @@ import java.util.concurrent.RejectedExecutionException;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.concurrent.atomic.AtomicReference;
-import org.hiero.consensus.concurrent.pool.StandardWorkGroup;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
