+++ platform-sdk/swirlds-base/build.gradle.kts
@@ -15,6 +15,7 @@ tasks.withType<JavaCompile>().configureEach {
 
 testModuleInfo {
     requires("com.swirlds.base.test.fixtures")
+    requires("org.hiero.base.utility.test.fixtures")
     requires("org.junit.jupiter.api")
     requires("org.assertj.core")
     requires("org.mockito")
