+++ platform-sdk/swirlds-base/src/test/java/com/swirlds/base/formatting/HorizontalAlignmentTests.java
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: Apache-2.0
-package com.swirlds.common.formatting;
+package com.swirlds.base.formatting;
 
 import static com.swirlds.base.formatting.HorizontalAlignment.ALIGNED_CENTER;
 import static com.swirlds.base.formatting.HorizontalAlignment.ALIGNED_LEFT;
