+++ platform-sdk/swirlds-base/src/test/java/com/swirlds/base/formatting/TextEffectTests.java
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: Apache-2.0
-package com.swirlds.common.formatting;
+package com.swirlds.base.formatting;
 
 import static com.swirlds.base.formatting.TextEffect.CROSSED_OUT;
 import static com.swirlds.base.formatting.TextEffect.RED;
@@ -9,7 +9,6 @@ import static com.swirlds.base.formatting.TextEffect.getPrintableTextLength;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-import com.swirlds.base.formatting.TextEffect;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
